requirements.txt
mariadb~=1.0.10
python-dotenv~=0.19.2
djangorestframework~=3.13.1
Django~=4.0.2